refactor: Move unused performance utilities to legacy
Move gradient-cache.ts, performance-utils.ts, and quality-manager.ts to legacy as they are only used by legacy components. Update import paths in legacy components and document moved files in legacy README.
