+++ legacy/lib/performance/gradient-cache.ts
