+++ legacy/lib/performance/performance-utils.ts
