+++ legacy/lib/performance/quality-manager.ts
